Model Changes, Logic Changes, Can Now Save!
The model has undergone significant revisions. I've taken out lesser
tables (Ball, Destination, Trap, etc) and just moved them up to the
main table named "Level". I've also added a new table for storing user
information, but have yet to do anything with it.

I'm now able to save levels created in the LevelBuilder to Core Data
and retrieve them. The issue is that I want to display them in a table
for a user to select from (see MyTableViewController), but the issue is
that when I try to set the label (after initialization), it does not
seem to want to update the view. I've tried StackOverflow and every
resource I can think of, but am so far out of luck. I may just scrap
the table all together and implement my own sort of organization logic.

Progress is coming along, Objective-C is an interesting language and
I'm having a lot of fun with this.